refactor(date): move getDate function to a new file
